separated logic in FileReader
