+++ src/main/java/service/FileReader.java
@@ -1,8 +1,7 @@
 package service;
 
 import java.util.List;
-import model.FruitTransaction;
 
 public interface FileReader {
-    List<FruitTransaction> read(String inputFileName);
+    List<String> read(String inputFileName);
 }
