+++ src/main/java/service/FruitTransactionService.java
@@ -1,0 +1,8 @@
+package service;
+
+import java.util.List;
+import model.FruitTransaction;
+
+public interface FruitTransactionService {
+    List<FruitTransaction> parseTransactions(List<String> transactions);
+}
